HSEARCH-3638 Move TestElasticsearchClient and dependencies to a separate utils module
So that it can be used in ITs from the documentation module, too.
