+++ documentation/pom.xml
@@ -62,6 +62,12 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>hibernate-search-util-internal-integrationtest-elasticsearch</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>hibernate-search-util-internal-integrationtest-orm</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
