+++ documentation/src/test/java/org/hibernate/search/documentation/testsupport/ElasticsearchBackendConfiguration.java
@@ -6,16 +6,30 @@
  */
 package org.hibernate.search.documentation.testsupport;
 
+import java.util.Optional;
+
 import org.hibernate.search.backend.elasticsearch.cfg.ElasticsearchBackendSettings;
 import org.hibernate.search.util.impl.integrationtest.common.TestConfigurationProvider;
 import org.hibernate.search.util.impl.integrationtest.common.rule.MappingSetupHelper;
+import org.hibernate.search.util.impl.integrationtest.elasticsearch.rule.TestElasticsearchClient;
+
+import org.junit.rules.TestRule;
 
 public class ElasticsearchBackendConfiguration extends AbstractDocumentationBackendConfiguration {
+
+	// This will be used in a later commit
+	protected final TestElasticsearchClient testElasticsearchClient = new TestElasticsearchClient();
+
 	@Override
 	public String toString() {
 		return "elasticsearch";
 	}
-	
+
+	@Override
+	public Optional<TestRule> getTestRule() {
+		return Optional.of( testElasticsearchClient );
+	}
+
 	@Override
 	public <C extends MappingSetupHelper<C, ?, ?>.AbstractSetupContext> C setupWithName(C setupContext,
 			String backendName, TestConfigurationProvider configurationProvider) {
