+++ util/internal/integrationtest/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/elasticsearch/dialect/Elasticsearch5TestDialect.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.elasticsearch.testsupport.dialect;
+package org.hibernate.search.util.impl.integrationtest.elasticsearch.dialect;
 
 import java.util.Arrays;
 import java.util.List;
@@ -13,11 +13,10 @@ import java.util.Optional;
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 @SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch60TestDialect implements ElasticsearchTestDialect {
+public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public boolean isEmptyMappingPossible() {
@@ -46,8 +45,16 @@ public class Elasticsearch60TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public void setTemplatePattern(JsonObject object, String pattern) {
-		JsonArray array = new JsonArray();
-		array.add( pattern );
-		object.add( "index_patterns", array );
+		object.addProperty( "template", pattern );
+	}
+
+	@Override
+	public boolean isGeoPointIndexNullAsPossible() {
+		return false;
+	}
+
+	@Override
+	public boolean worksFineWithStrictGraterThanRangedQueriesOnScaledFloatField() {
+		return false;
 	}
 }
