+++ util/internal/integrationtest/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/elasticsearch/dialect/Elasticsearch60TestDialect.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.elasticsearch.testsupport.dialect;
+package org.hibernate.search.util.impl.integrationtest.elasticsearch.dialect;
 
 import java.util.Arrays;
 import java.util.List;
@@ -17,7 +17,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 @SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch67TestDialect implements ElasticsearchTestDialect {
+public class Elasticsearch60TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public boolean isEmptyMappingPossible() {
@@ -36,7 +36,7 @@ public class Elasticsearch67TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public Boolean getIncludeTypeNameParameterForMappingApi() {
-		return true;
+		return null;
 	}
 
 	@Override
