+++ util/internal/integrationtest/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/elasticsearch/dialect/Elasticsearch7TestDialect.java
@@ -4,33 +4,33 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.elasticsearch.testsupport.dialect;
+package org.hibernate.search.util.impl.integrationtest.elasticsearch.dialect;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 import org.hibernate.search.backend.elasticsearch.client.impl.Paths;
 import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
+import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
-@SuppressWarnings("deprecation") // We use Paths.DOC on purpose
-public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
+public class Elasticsearch7TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public boolean isEmptyMappingPossible() {
-		return true;
+		return false;
 	}
 
 	@Override
 	public URLEncodedString getTypeKeywordForNonMappingApi() {
-		return Paths.DOC;
+		return Paths._DOC;
 	}
 
 	@Override
 	public Optional<URLEncodedString> getTypeNameForMappingApi() {
-		return Optional.of( Paths.DOC );
+		return Optional.empty();
 	}
 
 	@Override
@@ -40,21 +40,13 @@ public class Elasticsearch5TestDialect implements ElasticsearchTestDialect {
 
 	@Override
 	public List<String> getAllLocalDateDefaultMappingFormats() {
-		return Arrays.asList( "yyyy-MM-dd", "yyyyyyyyy-MM-dd" );
+		return Collections.singletonList( "uuuu-MM-dd" );
 	}
 
 	@Override
 	public void setTemplatePattern(JsonObject object, String pattern) {
-		object.addProperty( "template", pattern );
-	}
-
-	@Override
-	public boolean isGeoPointIndexNullAsPossible() {
-		return false;
-	}
-
-	@Override
-	public boolean worksFineWithStrictGraterThanRangedQueriesOnScaledFloatField() {
-		return false;
+		JsonArray array = new JsonArray();
+		array.add( pattern );
+		object.add( "index_patterns", array );
 	}
 }
