+++ util/internal/integrationtest/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/elasticsearch/dialect/ElasticsearchTestDialect.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.elasticsearch.testsupport.dialect;
+package org.hibernate.search.util.impl.integrationtest.elasticsearch.dialect;
 
 import java.util.List;
 import java.util.Optional;
