+++ util/internal/integrationtest/elasticsearch/src/main/resources/test-elasticsearch-client.properties
@@ -1,0 +1,8 @@
+hosts ${test.elasticsearch.host.url}
+username ${test.elasticsearch.host.username}
+password ${test.elasticsearch.host.password}
+aws.signing.enabled ${test.elasticsearch.host.aws.signing.enabled}
+aws.signing.access_key ${test.elasticsearch.host.aws.signing.access_key}
+aws.signing.secret_key ${test.elasticsearch.host.aws.signing.secret_key}
+aws.signing.region ${test.elasticsearch.host.aws.signing.region}
+log.json_pretty_printing true
