+++ util/internal/integrationtest/pom.xml
@@ -27,6 +27,7 @@
 
     <modules>
         <module>common</module>
+        <module>elasticsearch</module>
         <module>orm</module>
         <module>sharedresources</module>
     </modules>
